C4-38 fix
